Networking events to prefer features rather than events and features
If a feature exists with declared event name, that feature will be called directly
rather than via the event. The primary motivator was that the events were an unnecessary
complication in a complicated process that is prone to bugs.

The old way will still work for now, but will likely become deprecated soon.